Day 6 part 2 complete

+++ src/test/java/Day6Test.java
@@ -6,7 +6,7 @@ class Day6Test {
 
     @Test
     public void testDay6() {
-        int answer = new Day6().day6();
+        long answer = new Day6().day6();
         System.out.println("Answer total: " + answer);
     }
 }
